Implement version management page (in progress)
